# Conflicts:
#	anipy_cli/__init__.py
#	anipy_cli/cli.py
#	anipy_cli/config.py
#	anipy_cli/seasonal.py
#	setup.py

+++ anipy_cli/cli.py
@@ -19,20 +19,27 @@ from .misc import (
     clear_console,
     print_names,
     seasonal_options,
+    parsenum,
     mal_options,
     search_in_season_on_gogo,
 )
-from .player import mpv, dc_presence_connect
-from .misc import error, entry, options, clear_console, print_names, seasonal_options
-from .player import start_player, dc_presence_connect
+from .player import (
+    start_player,
+    dc_presence_connect,
+    create_mpv_controllable,
+    mpv_start_stream,
+)
 from .query import query
 from .arg_parser import parse_args
 from .colors import colors
 from .download import download
-from .config import config
+from .config import Config
+
+# Make colors work in windows CMD
+os.system("")
 
 rpc_client = None
-if config.dc_presence:
+if Config().dc_presence:
     rpc_client = dc_presence_connect()
 
 
@@ -53,8 +60,15 @@ def default_cli(quality, player):
     url_class.stream_url()
     show_entry = url_class.get_entry()
 
-    sub_proc = start_player(show_entry, rpc_client, player)
-    menu(show_entry, options, sub_proc, quality, player).print_and_input()
+    mpv = None
+    sub_proc = None
+    if Config().reuse_mpv_window and not player and Config().player_path == "mpv":
+        mpv = create_mpv_controllable()
+        mpv = mpv_start_stream(show_entry, mpv, rpc_client)
+    else:
+        sub_proc = start_player(show_entry, rpc_client, player)
+
+    menu(show_entry, options, sub_proc, quality, player, mpv).print_and_input()
 
 
 def download_cli(quality, ffmpeg, no_season_search):
@@ -68,7 +82,7 @@ def download_cli(quality, ffmpeg, no_season_search):
         colors.GREEN
         + "Downloads are stored in: "
         + colors.END
-        + str(config.download_folder_path)
+        + str(Config().download_folder_path)
     )
 
     show_entry = entry()
@@ -120,7 +134,7 @@ def download_cli(quality, ffmpeg, no_season_search):
         show_entry = ent["show_entry"]
         ep_list = ent["ep_list"]
         for i in ep_list:
-            show_entry.ep = int(i)
+            show_entry.ep = parsenum(i)
             show_entry.embed_url = ""
             ep_class = epHandler(show_entry)
             show_entry = ep_class.gen_eplink()
@@ -414,12 +428,13 @@ class menu:
     a subprocess instance returned by misc.start_player().
     """
 
-    def __init__(self, entry, opts, sub_proc, quality, player) -> None:
+    def __init__(self, entry, opts, sub_proc, quality, player, mpv=None) -> None:
         self.entry = entry
         self.options = opts
         self.sub_proc = sub_proc
         self.quality = quality
         self.player = player
+        self.mpv = mpv
 
     def print_opts(self):
         for i in self.options:
@@ -440,6 +455,9 @@ class menu:
         self.take_input()
 
     def kill_player(self):
+        if self.mpv:
+            return
+
         self.sub_proc.kill()
 
     def start_ep(self):
@@ -447,7 +465,10 @@ class menu:
         url_class = videourl(self.entry, self.quality)
         url_class.stream_url()
         self.entry = url_class.get_entry()
-        self.sub_proc = start_player(self.entry, rpc_client, self.player)
+        if self.mpv:
+            self.mpv = mpv_start_stream(self.entry, self.mpv, rpc_client)
+        else:
+            self.sub_proc = start_player(self.entry, rpc_client, self.player)
 
     def take_input(self):
         while True:
@@ -526,8 +547,9 @@ class menu:
         print(f"Quality: {self.entry.quality}")
 
     def quit(self):
-
         self.kill_player()
+        if self.mpv:
+            self.mpv.terminate()
         sys.exit(0)
 
 
@@ -986,6 +1008,7 @@ class MALCli:
 
 
 def main():
+
     args = parse_args()
 
     player = None
@@ -998,7 +1021,7 @@ def main():
 
     if args.delete:
         try:
-            config.history_file_path.unlink()
+            Config().history_file_path.unlink()
             print(colors.RED + "Done")
         except FileNotFoundError:
             error("no history file found")
@@ -1023,7 +1046,7 @@ def main():
         history_cli(args.quality, player)
 
     elif args.config:
-        print(os.path.realpath(__file__).replace("cli.py", "config.py"))
+        print(Config()._config_file)
 
     elif args.mal:
         mal_cli(
